Extend SFXAPIError to include token and endpoint
This change extends the existing SFXAPIError struct by adding two extra
fields 'token' and 'endpoint'. This allows for custom error
handlers that can behave differently depending on the token or endpoint
associated with the error. In the AsyncMultiTokenSink case, this error
handler is only called when the underlying client can't successfully
call the ingestion endpoint after the configured number of retries.

An example of where this could be useful:

The token for SignalFx ingestion is being throttled, so report the
error directly to cloudwatch so we can alert on it, with the token
and endpoint affected as dimensions for more targeted alerting.